fix(entity): corriger mismatches Hibernate validate — imf_id BIGINT, attempts_used SMALLINT, uuid_mobile UUID
- ClientScore.imfId : Integer → Long (imf_id BIGINT en DB, V23/V29)
- ClientScoreRepository + ScoringResultConsumer : cast (long) pour compatibilité Avro int
- OtpCode.attemptsUsed : int → short (attempts_used SMALLINT en DB, V30)
- AuthService : cast (short) sur l'incrémentation pour éviter l'erreur de compilation
- CollecteEpargne.uuidMobile : ajout columnDefinition="uuid" pour cohérence Hibernate 6

+++ backend/src/main/java/cm/imf/pipeline/entity/ClientScore.java
@@ -30,7 +30,7 @@ public class ClientScore {
     private String clientIdExterne;
 
     @Column(name = "imf_id", nullable = false)
-    private Integer imfId;
+    private Long imfId;
 
     @Column(name = "score_mcrs", nullable = false)
     private Double scoreMcrs;
